rag_app/requirements.txt
streamlit==1.32.0
google-cloud-vision==3.7.0
google-cloud-aiplatform==1.42.0
faiss-cpu==1.7.4
numpy
Pillow
python-dotenv